clipboard-apis/async-write-blobtext-read-text-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Async Clipboard write (Blob/text) -> readText tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  const inputMessage = "Clipboard write (Blob/text) -> readText test データ";
  const input = new Blob([inputMessage], {type: 'text/plain'});

  await navigator.clipboard.write(input);
  const output = await navigator.clipboard.readText();

  assert_equals(output, inputMessage);
}, "Verify write and read clipboard (DataTransfer/text)");
</script>
Note: This is a manual test because it writes/reads to the shared system
clipboard and thus cannot be run async with other tests that might interact
with the clipboard.
